backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
bcrypt==4.0.1
pyjwt==2.8.0
python-dotenv==1.0.0
pillow==10.0.1
email-validator==2.0.0
werkzeug==2.3.7
bson==0.5.10
